requirements.txt
pandas
matplotlib
jupyter
ipykernel
python-dotenv
seaborn